Import live.2009.02.23.tar.gz
2009.02.13
- Changed calls to the "Locale()" constructor to use an explicit variable name (rather than an anonymous
  declaration), because some compilers weren't handling the anonymous declaration properly.
  (Thanks to Steven Kramer for noticing this.)
- Improved the "RTSPServer" definition and implementation to make it easier to subclass the
  "RTSPServer::RTSPClientSession" class (e.g., to add a custom implementation of the "SET_PARAMETER" command).
  Also added an iterator class for iterating over the "RTSPClientSession" objects within a "RTSPServer".
- Make "MediaSubsession::parseSDPAttribute_framerate()" a little more robust.
  (Thanks to Shawn Van Every for the suggestion.)
- Improved "BasicUsageEnvironment0::reportBackgroundError()" (thanks to Matt Schuckmann)
- Fixed a bug in the "AVIFileSink" code (thanks to Pramod Bhadwat for reporting this).

2009.02.23
- Changed the way in which 'trick mode' transport streams are generated, so that each
  video I-frame is included in the output Transport Stream no more than once.
  This means that now - for low scale values - the output frame rate, and bit rate,
  will be les than that of the original file.  (Beforehand, we maintained the same
  frame rate, which meant that - because each output frame is an I-frame - the output
  stream's bit rate was always significantly larger than the original.  This caused
  problems for some clients.)
- Changed the 'client liveness' checking code in the "RTSPServer" implementation so that it deletes the
  "RTSPClientSession" object even for multicast streams.
- Made a minor fix to an error report in "RTSPClient".  (Thanks to Igor Milavec.)

+++ testProgs/playCommon.cpp
@@ -1263,7 +1263,7 @@ void checkForPacketArrival(void* /*clientData*/) {
     struct timeval timeNow;
     gettimeofday(&timeNow, NULL);
 	char timestampStr[100];
-	sprintf(timestampStr, "%ld%03ld", timeNow.tv_sec, timeNow.tv_usec/1000);
+	sprintf(timestampStr, "%ld%03ld", timeNow.tv_sec, (long)(timeNow.tv_usec/1000));
     *env << (syncStreams ? "Synchronized d" : "D")
 		<< "ata packets have begun arriving [" << timestampStr << "]\007\n";
     return;
